requirements.txt
# requirements.txt
django-cms>=4.0.0
django-mcp-server>=0.1.0
django-filer>=3.0.0